requirements.txt
fastapi==0.115.12
uvicorn==0.33.0
numpy==1.24.4
python-multipart==0.0.20
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
ultralytics==8.3.153
pytorch-nlp==0.5.0
flashlight==0.1.1
flashlight-text==0.0.7

--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision
torchaudio
